implement complete update and delete functionality for user galleries

+++ src/pages/GalleryPage/GalleryPage.jsx
@@ -6,18 +6,6 @@ import { useState, useEffect, useRef } from "react";
 import * as apiService from "../../utilities/artsy-api-service";
 import { getUser } from "../../utilities/users-service.js";
 
-const defaultSlides = [
-	{
-		url: "https://d32dm0rphc51dk.cloudfront.net/m_l6G2TnBq7tM9TVOHp_Fw/large.jpg",
-		title: "Good Pic",
-	},
-	{
-		url: "https://d32dm0rphc51dk.cloudfront.net/MoR7cUheG5Iv4cM1HeR3CQ/large.jpg",
-		title: "Great Pic",
-	},
-	{ url: "https://i.imgur.com/JI8GVlK.jpeg", title: "Best Cat" },
-];
-
 export default function GalleryPage() {
 	const [artworks, setArtworks] = useState([]);
 	const loadingSlides = useRef(false);
@@ -29,19 +17,31 @@ export default function GalleryPage() {
 		margin: "0 auto",
 	};
 
+	var currentArtwork = "";
+
+	const setCurrentArtwork = (artwork) => {
+		currentArtwork = artwork;
+	};
+
 	async function getGallerySlides() {
 		gotSlides.current = false;
 		let user = getUser()._id;
 		let result = await apiService.getGalleryArt(user);
 		setArtworks(result);
+		setCurrentArtwork(artworks[0]);
 		gotSlides.current = true;
-		// use the above to look at the result from the database and format the images in the components below
 	}
 
-	async function removeArtworkfromGallery() {
-		let user = getUser()._id;
-		let artwork_id = "0";
-		await apiService.removeGalleryArt(user, artwork_id);
+	async function removeArtworkfromGallery(event) {
+		if (currentArtwork) {
+			let user = getUser()._id;
+			let artwork_id = currentArtwork.artwork_id;
+			await apiService.removeGalleryArt(user, artwork_id);
+		}
+	}
+
+	async function handleSubmit(event) {
+		await removeArtworkfromGallery();
 		await getGallerySlides();
 	}
 
@@ -56,7 +56,10 @@ export default function GalleryPage() {
 			<div className="background-wrapper">
 				{gotSlides ? (
 					<div style={containerStyles}>
-						<ImageSlider artworks={artworks} />
+						<ImageSlider
+							artworks={artworks}
+							setcurrentArtwork={setCurrentArtwork}
+						/>
 					</div>
 				) : (
 					<div>
@@ -65,8 +68,7 @@ export default function GalleryPage() {
 				)}
 			</div>
 			<div className="galleryButtons">
-				<button>Show Similar Art</button>
-				<button onClick={removeArtworkfromGallery}>Remove From My Gallery</button>
+				<button onClick={handleSubmit}>Remove From My Gallery</button>
 			</div>
 			{/* <BenchFooter /> */}
 		</>
